bug解决

+++ src/main/java/com/example/demo/service/OrderServiceImpl.java
@@ -7,8 +7,12 @@ import com.example.demo.entity.OrderBus;
 import com.example.demo.util.OperatorInfoUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
 import java.util.List;
+
+import static com.example.demo.constant.OrderStatus.ACTIVE;
+import static com.example.demo.constant.OrderStatus.ORDER;
 
 /**
  * OrderServiceImpl
@@ -27,6 +31,12 @@ public class OrderServiceImpl implements IOrderService {
     @Override
     public void saveOrder(Order order, List<OrderBus> busList) {
         OperatorInfoUtil.fillOperateInfo(order);
+        order.setStatus(ACTIVE.getName());
+        if(CollectionUtils.isEmpty(busList)){
+            order.setStatus(ORDER.getName());
+            orderDAO.insert(order);
+            return;
+        }
         orderDAO.insert(order);
         for (OrderBus bus : busList) {
             OperatorInfoUtil.fillOperateInfo(bus);
@@ -38,6 +48,12 @@ public class OrderServiceImpl implements IOrderService {
     @Override
     public void editOrder(Order order, List<OrderBus> busList) {
         OperatorInfoUtil.fillUpdateInfo(order);
+        order.setStatus(ACTIVE.getName());
+        if(CollectionUtils.isEmpty(busList)){
+            order.setStatus(ORDER.getName());
+            orderDAO.updateByPrimaryKey(order);
+            return;
+        }
         orderDAO.updateByPrimaryKey(order);
         orderBusDAO.deleteByOrdId(order.getOrdId());
         for (OrderBus bus : busList) {
